refactor(centre-aide): Callout/Accordeon à la racine kpilote-ui, helpers d'URL dans kpilote-shared
