+++ packages/kpilote-ui/src/centre-aide/index.ts
@@ -1,15 +1,7 @@
 export { RenduContenuCentreAide, classesRenduBase } from './RenduContenuCentreAide'
-export { Callout } from './Callout'
-export { Accordeon } from './Accordeon'
 export { ImageCentreAide } from './ImageCentreAide'
 export { VideoCentreAide } from './VideoCentreAide'
 export { IconeCentreAide } from './IconeCentreAide'
 export { registreBlocs } from './registreBlocs'
 export { registreIcones, TYPES_ICONE } from './registreIcones'
-export {
-  COULEURS_CALLOUT,
-  type BlocCentreAideType,
-  type CalloutColor,
-  type DescripteurBloc,
-  type RegistreBlocs,
-} from './types'
+export { type BlocCentreAideType, type DescripteurBloc, type RegistreBlocs } from './types'
