+++ packages/kpilote-ui/src/centre-aide/types.ts
@@ -2,10 +2,6 @@ import type { ReactNode } from 'react'
 
 export type BlocCentreAideType = 'callout' | 'accordion-item' | 'image' | 'icone' | 'video'
 
-export type CalloutColor = 'info' | 'success' | 'warning' | 'error'
-
-export const COULEURS_CALLOUT: readonly CalloutColor[] = ['info', 'success', 'warning', 'error']
-
 export type RendreEnfants = (element: Element) => ReactNode
 
 export interface DescripteurBloc {
